requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via pytest
dj-database-url==2.1.0
    # via -r requirements.in
django==5.0.6
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-redis
    #   djangorestframework
django-redis==5.4.0
    # via -r requirements.in
djangorestframework==3.15.1
    # via -r requirements.in
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
packaging==24.0
    # via pytest
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
redis==5.0.4
    # via django-redis
requests==2.32.2
    # via -r requirements.in
sqlparse==0.5.0
    # via django
typing-extensions==4.11.0
    # via dj-database-url
tzdata==2024.1
    # via django
urllib3==2.2.1
    # via requests
